requirements.txt
# --- CORE ---
langchain==0.3.7
langchain-core==0.3.19
langchain-community==0.3.7

# --- OpenAI integration (compatible with langchain 0.3) ---
langchain-openai==0.2.2

# --- Vector DB ---
langchain-chroma==0.1.4
chromadb==0.5.3

# --- Utilities ---
python-dotenv==1.0.1
pandas==2.2.2
requests==2.31.0
tqdm==4.66.4
openpyxl
==3.1.3

#dashboarding
streamlit==1.39.0
plotly==5.24.1
altair==5.3.0


